% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTGSA-package.R, R/CTGSA.R
\name{CTGSA}
\alias{CTGSA}
\title{CTGSA}
\usage{
CTGSA(x, y, out.type = c("stat", "pvalue"),
  percentile.type = c("case-control", "all", "user-defined"),
  percentile.user = NULL, nperm = NULL)
}
\arguments{
\item{x}{an input matrix of dimension nobs x nvars.}

\item{y}{a response variable where y=1 is the case and y=0 is control.}

\item{out.type}{character of output type of statistic('stat') and p-value('pvalue').}

\item{percentile.type}{three threshold calculating types are supported as 'case-control', 'all', and 'user-defined'. To choose the threshold, 'case-control' computes the percentile with the first singular value for case and control. Percentile.type 'all' calculates the percentile with whole samples. The percentile is given by user in 'user-defined', and the threshold is then calculated for case and control. The default is case-control.}

\item{percentile.user}{a numeric value of percentile in [0, 1]. To specify the percentile, give the percentile.type 'user-defined', and the integer value taken here will be used.}

\item{nperm}{number of permuation - default is 1000.}
}
\value{
A list of a statistic, an optional pvalue and permuted statistics, and arguments used.
}
\description{
CTGSA

CTGSA : Covariance Thresholding for Gene Set Analysis of Genetic Pathway Data
}
\details{
See the reference below for more information.
}
\examples{

x <- replicate(5, rnorm(100, 0, 1))
y <- rep(0:1, each=50)
CTGSA( x = x, y = y, out.type="stat", percentile.type = "all")


require(mnormt)
x <- rbind(rmnorm(50, rep(0, 5), diag(5) ),
           rmnorm(50, rep(0, 5), outer(1:5, 1:5, function(i, j) 0.6^abs(i-j) ) ))
y <- rep(0:1, each=50)
CTGSA(x=x, y=y, out.type="pvalue", percentile.type = "case-control", nperm=1000)

}
